Pie Chart finish, up to this moment only support int value, modify multiChart UI to add one more button, it's okay to modify it again.

+++ src/core/comp3111/PieChartObj.java
@@ -19,8 +19,8 @@ public class PieChartObj extends GeneralChart {
 		this.chartName = "Chart " + (dt.getStoredChart().size() + 1) + " [Type = Pie] [Items: " + stringItemsName + ", Quantity: " + numItemsName + "]";
 		pieChart.setTitle(stringItemsName + " Distriubution");
 		for(int i = 0; i < number.getSize(); i++) {
-			pieChartData.add(new PieChart.Data((String)textColumn.getData()[i], (double)numColumn.getData()[i]));
-			System.out.println("Added " + ( i + 1 ) + " data. [ Items = " + (String)textColumn.getData()[i] + " ] [ Quantity = " + (double)numColumn.getData()[i] + " ]");
+			pieChartData.add(new PieChart.Data((String)textColumn.getData()[i], (int) numColumn.getData()[i]));
+			System.out.println("Added " + ( i + 1 ) + " data. [ Items = " + (String)textColumn.getData()[i] + " ] [ Quantity = " + (int)numColumn.getData()[i] + " ]");
 		}
 		this.chart = pieChart;
 		System.out.println("[ NEW Line Chart : " + this.chartName + " created ]");
